refactor: type, utils 폴더 분리 및 배포 package.json 수정
