+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-component-layout",
   "private": false,
-  "version": "1.1.0",
+  "version": "1.2.8",
   "homepage": "https://nlom0218.github.io/layout-component",
   "keywords": [
     "react",
@@ -9,12 +9,12 @@
     "react-component-layout",
     "component-layout"
   ],
-  "main": "dist/index.umd.cjs",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "./dist/index.umd.cjs",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/index.js",
+      "import": "./dist/index.mjs",
       "require": "./dist/index.umd.cjs"
     }
   },
