+++ src/lib/index.tsx
@@ -1,3 +1,4 @@
 export { default as Container } from './Container/Container';
 export { default as Grid } from './Grid/Grid';
 export { default as Flex } from './Flex/Flex';
+export { default as Tabs } from './Tabs/Tabs';
