--- src/lib/type.ts
@@ -1,5 +1,0 @@
-export type ResponsiveStyleType<T> = Partial<{
-  desktop: T;
-  tablet: T | [number, T];
-  mobile: T | [number, T];
-}>;
