+++ src/lib/types/type.ts
@@ -1,5 +1,11 @@
 import { ReactNode } from 'react';
 
+export type ResponsiveStyleType<T> = Partial<{
+  desktop: T;
+  tablet: T | [number, T];
+  mobile: T | [number, T];
+}>;
+
 export type Label = string | number;
 
 export type PanelList = {
